+++ ostsdk/src/main/java/com/ost/walletsdk/network/OstJsonApiCallback.java
@@ -1,7 +1,7 @@
 package com.ost.walletsdk.network;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import com.ost.walletsdk.annotations.NonNull;
+import com.ost.walletsdk.annotations.Nullable;
 
 import com.ost.walletsdk.workflows.errors.OstError;