+++ ostsdk/src/main/java/com/ost/walletsdk/ui/resetpin/ResetPinView.java
@@ -24,4 +24,6 @@ interface ResetPinView extends BaseView {
     void gotoDashboard(String workflowId);
 
     void showPinErrorDialog();
+
+    void onPostAuthentication();
 }