+++ ostsdk/src/main/java/com/ost/walletsdk/ui/workflow/OstLoaderCompletionDelegate.java
@@ -1,0 +1,5 @@
+package com.ost.walletsdk.ui.workflow;
+
+public interface OstLoaderCompletionDelegate {
+    void dismissWorkflow();
+}